# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CPNS & School Tutoring Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Packages
...
  /quizzes/{id}/start:
    post:
      operationId: startQuizAttempt
      tags: [attempts]
      summary: Start a quiz attempt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Attempt started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /tryouts/{id}/start:
    post:
      operationId: startTryoutAttempt
      tags: [attempts]
      summary: Start a tryout attempt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Attempt started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attempt"
  /attempts/{id}/submit:
    post:
      operationId: submitAttempt
      tags: [attempts]
      summary: Submit answers for an attempt
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttemptSubmitInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResult"
  /attempts/{id}:
    get:
      operationId: getAttempt
      tags: [attempts]
      summary: Get attempt result detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttemptResult"
  /attempts:
    get:
      operationId: listAttempts
      tags: [attempts]
      summary: List my attempts (student)
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: quizId
          in: query
          schema:
            type: integer
        - name: tryoutId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attempt"

  # Admin dashboard
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/revenue:
    get:
      operationId: getRevenueChart
      tags: [admin]
      summary: Get monthly revenue data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueData"

  # Student dashboard
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, tutor, student]
        avatar:
          type: ["string", "null"]
        targetInstitution:
          type: ["string", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
    UserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Package
...
    Attempt:
      type: object
      required: [id, userId, type, startedAt, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        quizId:
          type: ["integer", "null"]
        tryoutId:
          type: ["integer", "null"]
        type:
          type: string
          enum: [quiz, tryout]
        score:
          type: ["number", "null"]
        startedAt:
          type: string
        finishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        quizTitle:
          type: ["string", "null"]
        tryoutTitle:
          type: ["string", "null"]
    AttemptSubmitInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: object
          additionalProperties:
            type: string
    AttemptResult:
      type: object
      required: [id, userId, type, score, totalQuestions, correctAnswers, startedAt, finishedAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        quizId:
          type: ["integer", "null"]
        tryoutId:
          type: ["integer", "null"]
        type:
          type: string
        score:
          type: number
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        twkScore:
          type: ["number", "null"]
        tiuScore:
          type: ["number", "null"]
        tkpScore:
          type: ["number", "null"]
        passed:
          type: ["boolean", "null"]
        rank:
          type: ["integer", "null"]
        totalParticipants:
          type: ["integer", "null"]
        startedAt:
          type: string
        finishedAt:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerDetail"
    AnswerDetail:
      type: object
      required: [questionId, questionText, userAnswer, correctAnswer, isCorrect]
      properties:
        questionId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        optionE:
          type: string
        userAnswer:
          type: ["string", "null"]
        correctAnswer:
          type: string
        isCorrect:
          type: boolean
        explanation:
          type: ["string", "null"]
        category:
          type: ["string", "null"]

    # Admin stats
    AdminStats:
      type: object
      required: [totalStudents, totalRevenue, pendingOrders, totalPackages, activeEnrollments]
      properties:
        totalStudents:
          type: integer
        totalRevenue:
          type: integer
        pendingOrders:
          type: integer
        totalPackages:
          type: integer
        activeEnrollments:
          type: integer
        newStudentsThisMonth:
          type: integer
        revenueThisMonth:
          type: integer
    RevenueData:
      type: object
      required: [month, revenue, orders]
      properties:
        month:
          type: string
        revenue:
          type: integer
        orders:
          type: integer

    # Student dashboard
...